feat: Adiciona fluxo para controle de contagem de votos e ranking pelo redis e comunicação em tempo real com websocket

+++ src/routes/vote-on-poll.ts
@@ -2,6 +2,8 @@ import { FastifyInstance } from "fastify";
 import { randomUUID } from "crypto";
 import { prisma } from "../lib/prisma";
 import z from "zod";
+import { redis } from "../lib/redis";
+import { voting } from "../utils/voting-pub-sub";
 
 export async function voteOnPoll(app: FastifyInstance) {
   app.post("/polls/:pollId/votes", async (request, reply) => {
@@ -37,11 +39,23 @@ export async function voteOnPoll(app: FastifyInstance) {
           .send({ message: "You already voted on this poll!" });
       }
 
-      await prisma.vote.delete({
-        where: {
-          id: userPreviousVoteOnPoll?.id,
-        },
-      });
+      if (userPreviousVoteOnPoll) {
+        await prisma.vote.delete({
+          where: {
+            id: userPreviousVoteOnPoll.id,
+          },
+        });
+
+        const votes = await redis.zincrby(
+          pollId,
+          -1,
+          userPreviousVoteOnPoll.pollOptionId
+        );
+        voting.publish(pollId, {
+          pollOptionId: userPreviousVoteOnPoll.pollOptionId,
+          votes: Number(votes),
+        });
+      }
     }
 
     if (!sessionId) {
@@ -60,6 +74,12 @@ export async function voteOnPoll(app: FastifyInstance) {
         pollId,
         pollOptionId,
       },
+    });
+
+    const votes = await redis.zincrby(pollId, 1, pollOptionId);
+    voting.publish(pollId, {
+      pollOptionId,
+      votes: Number(votes),
     });
 
     return reply.status(201).send();
